requirements.txt
# FastAPI and server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
jinja2==3.1.3

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Configuration
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Database
pymongo==4.6.1

# LLM
groq==0.4.1

# PDF Processing
PyPDF2==3.0.1
pypdf==3.17.4

# Embeddings and Vector Search
sentence-transformers==2.3.1
faiss-cpu==1.7.4
numpy==1.24.3

# Text Processing (optional but useful)
langchain==0.1.0
langchain-community==0.0.13

# HTTP client
httpx==0.26.0
requests==2.31.0